requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
requests>=2.31.0
python-decouple>=3.8
django-cors-headers>=4.3.0
geopy>=2.4.1
